% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot.R
\name{uniprot_context}
\alias{uniprot_context}
\title{Generate LLM-friendly context for the rest.uniprot.org package}
\usage{
uniprot_context()
}
\value{
Character string (invisibly), also printed
}
\description{
Generate LLM-friendly context for the rest.uniprot.org package
}
